Renames NewGameBox to GameBox and deletes the old one
