+++ src/components/GameBox.vue
@@ -1,533 +1,611 @@
 <template>
-  <div
-    ref="boxElement"
-    class="box reset-outer-spacing game-box hover-grow"
+  <article
+    class="game-box"
     :class="{
-      'is-vertical-video': isMobile,
-      'is-horizontal-video': !isMobile,
-      'is-init-video-only': isInitVideoOnly,
-      'is-details-at-right': isDetailsAtRight,
-      'is-details-at-left': isDetailsAtLeft,
+      'is-vertical-video': isVerticalVideo,
+      'is-horizontal-video': !isVerticalVideo,
+      'is-selected': isSelected,
+      'is-not-selected': !isSelected,
     }"
+    @click.prevent.stop="toggleGameSelection"
   >
-    <div class="box-content">
+    <div class="screen-overlay" />
+    <div class="game-box-content">
       <div class="game-box-middle">
-        <div class="video-inner-header">
-          <h2 class="game-title title title-simple">
+        <header class="video-header">
+          <h3 class="game-title title page-section-subtitle">
             {{ title }}
-          </h2>
-          <p v-if="isPersonal" class="video-inner-tag">
-            personal prototype
-          </p>
-          <p v-if="isFreelance" class="video-inner-tag">
-            as freelancer
-          </p>
-        </div>
-        <AutoplayVideo
-          class="game-video"
-          :src="videoSrc"
-          :width="sizeVars.videoWidth"
-          :height="sizeVars.videoHeight"
-          :poster="links.poster"
-        />
-        <div class="game-links">
-          <GameLink
-            v-if="links.steam"
-            title="Steam"
-            :href="links.steam"
-            icon-classes="fa-brands fa-steam-symbol"
+          </h3>
+        </header>
+        <section class="game-preview-section">
+          <h4 class="title">
+            Preview
+          </h4>
+          <AutoplayVideo
+            class="game-preview"
+            :src="videoSrc"
+            :poster="links.poster"
           />
-          <GameLink
-            v-if="links.googlePlay"
-            title="Store"
-            :href="links.googlePlay"
-            icon-classes="fa-brands fa-google-play"
-          />
-          <GameLink
-            v-if="links.trailer"
-            title="Trailer"
-            :href="links.trailer"
-            icon-classes="fa-solid fa-play"
-          />
-          <GameLink
-            v-if="links.gameplay"
-            title="Gameplay"
-            :href="links.gameplay"
-            icon-classes="fa-solid fa-gamepad"
-          />
-        </div>
+          <button
+            class="preview-info-button button is-rounded is-link is-outlined is-small"
+            @click.prevent
+          >
+            <span class="icon">
+              <i class="fa-solid fa-info" title="Show {{ title }} information" />
+            </span>
+          </button>
+        </section>
+        <nav class="game-links-section">
+          <h4 class="title">
+            Links
+          </h4>
+          <div class="game-links">
+            <div
+              v-if="links.steam"
+              class="game-link-wrapper"
+            >
+              <GameLink
+                title="Steam"
+                :href="links.steam"
+                icon-classes="fa-brands fa-steam-symbol"
+                @click.stop
+              />
+            </div>
+            <div
+              v-if="links.googlePlay"
+              class="game-link-wrapper"
+            >
+              <GameLink
+                title="Store"
+                :href="links.googlePlay"
+                icon-classes="fa-brands fa-google-play"
+                @click.stop
+              />
+            </div>
+            <div
+              v-if="links.trailer"
+              class="game-link-wrapper"
+            >
+              <GameLink
+                title="Trailer"
+                :href="links.trailer"
+                icon-classes="fa-solid fa-play"
+                @click.stop
+              />
+            </div>
+            <div
+              v-if="links.gameplay"
+              class="game-link-wrapper"
+            >
+              <GameLink
+                title="Gameplay"
+                :href="links.gameplay"
+                icon-classes="fa-solid fa-gamepad"
+                @click.stop
+              />
+            </div>
+          </div>
+        </nav>
       </div>
-      <div class="game-box-bottom">
-        <div class="details-columns">
-          <div class="details-column highlights-column">
-            <div class="game-highlights">
-              <h3 class="title title-simple is-5">
-                Highlights
-              </h3>
-              <BulletList
-                icon-classes="fa-solid fa-star"
-                :hover-color="starColor"
+      <GameToggableContent
+        ref="toggableContent"
+        class="game-box-bottom"
+        :game-box="$el"
+        :show-label="`${title} Info`"
+        :hide-label="`${title} Info`"
+        @toggled="refreshGameSelection"
+      >
+        <div class="game-sections">
+          <section class="game-section game-highlights">
+            <h4 class="title game-section-title">
+              Highlights
+            </h4>
+            <BulletList
+              icon-classes="fa-solid fa-star"
+            >
+              <BulletListItem v-for="topic in highlights" :key="topic">
+                {{ topic }}
+              </BulletListItem>
+            </BulletList>
+          </section>
+          <section
+            v-if="integrations?.length > 0"
+            class="game-section game-integrations"
+          >
+            <h4 class="title game-section-title">
+              Integrations
+            </h4>
+            <ul class="inline-list">
+              <li
+                v-for="integration in integrations"
+                :key="integration"
+                class="list-item"
               >
-                <BulletListItem v-for="topic in highlights" :key="topic">
-                  {{ topic }}
-                </BulletListItem>
-              </BulletList>
+                {{ integration }}
+              </li>
+            </ul>
+          </section>
+          <section class="game-section game-details">
+            <div class="game-details-wrapper">
+              <h4 class="title game-section-title">
+                Details
+              </h4>
+              <section class="game-details-section">
+                <h5 class="game-details-section-title">
+                  Made with
+                </h5>
+                <p class="game-details-section-content">
+                  {{ engine }}
+                </p>
+              </section>
+              <section class="game-details-section">
+                <h5 class="game-details-section-title">
+                  Platforms
+                </h5>
+                <ul class="inline-list game-details-section-content">
+                  <li
+                    v-for="platform in platforms"
+                    :key="platform"
+                    class="list-item"
+                  >
+                    {{ platform }}
+                  </li>
+                </ul>
+              </section>
+              <section class="game-details-section">
+                <h5 class="game-details-section-title">
+                  Commercial Context
+                </h5>
+                <p class="game-details-section-content">
+                  {{ commercialLabel }}
+                </p>
+              </section>
             </div>
-          </div>
-          <div class="details-column bullets-column">
-            <div v-if="integrations.length > 0" class="game-integrations">
-              <h3 class="title title-simple is-5">
-                Integrations
-              </h3>
-              <ul class="inline-list">
-                <li
-                  v-for="integration in integrations"
-                  :key="integration"
-                  class="list-item"
-                >
-                  {{ integration }}
-                </li>
-              </ul>
-            </div>
-            <div class="game-specs">
-              <h3 class="title title-simple is-5">
-                Specs
-              </h3>
-              <p>
-                Platforms:
-                <span
-                  v-for="platform in platforms"
-                  :key="platform"
-                  class="list-item"
-                >
-                  {{ platform }}
-                </span>
-              </p>
-              <p>
-                Made with: {{ engine }}
-              </p>
-            </div>
-          </div>
+          </section>
         </div>
-      </div>
+      </GameToggableContent>
     </div>
-  </div>
+  </article>
 </template>
 
 <script setup>
-import {
-    ref,
-    reactive,
-    onBeforeMount,
-    onMounted,
-    onUnmounted,
-    watch,
-} from 'vue';
+import { ref } from 'vue';
 import AutoplayVideo from './AutoplayVideo.vue';
 import BulletList from './BulletList.vue';
 import BulletListItem from './BulletListItem.vue';
 import GameLink from './GameLink.vue';
-import bulmaConstants from '../js/bulma-constants';
-import { px, preventSequentialCalls } from '../js/utils';
+import GameToggableContent from './GameToggableContent.vue';
+import { px } from '../js/utils';
 
 const BASE_HORIZONTAL_VSIZE = { width: 640, height: 360 };
 const BASE_VERTICAL_VSIZE = { width: 270, height: 480 };
+const COMMERCIAL_LABELS = {
+    'self-published': 'Self-Published Indie',
+    'freelance-job': 'Freelance Job',
+    'royalty-earning': 'Royalty Earning',
+    personal: 'Personal Demo',
+    prototype: 'Personal Prototype',
+};
+
+const { game } = defineProps({
+    game: { type: Object, required: true },
+});
+const isSelected = ref(false);
+const toggableContent = ref(null);
 
 const {
-    isMobile,
-    isInitVideoOnly,
-    minScale,
-    lerpScaleMinScreen,
-    lerpScaleMaxScreen,
-} = defineProps({
-    title: { type: String, required: true },
-    videoSrc: { type: String, required: true },
-    highlights: { type: Array, required: true },
-    integrations: { type: Array, required: true, default: () => [] },
-    platforms: { type: Array, required: true },
-    engine: { type: String, required: true },
-    links: { type: Object, required: true },
-    isPersonal: { type: Boolean, default: false },
-    isFreelance: { type: Boolean, default: false },
-    isMobile: { type: Boolean, default: false },
-    isInitVideoOnly: { type: Boolean, default: false },
-    minScale: { type: Number, default: 1 },
-    lerpScaleMinScreen: { type: Number, default: 0 },
-    lerpScaleMaxScreen: { type: Number, default: 0 },
-});
+    shortTitle: title,
+    isMobile: isVerticalVideo,
+    platforms,
+    links,
+    engine,
+} = game;
+const {
+    previewLow: videoSrc,
+} = links;
+const {
+    technical: highlights,
+    integrations,
+} = game.bullets;
+const commercialLabel = COMMERCIAL_LABELS[game.commercial];
 
-const baseVideoSize = isMobile
+const baseVideoSize = isVerticalVideo
     ? BASE_VERTICAL_VSIZE
     : BASE_HORIZONTAL_VSIZE;
 
-const boxElement = ref(null);
-const isDetailsAtRight = ref(true);
-const isDetailsAtLeft = ref(false);
-const sizeVars = reactive({
-    videoWidth: 0,
-    videoHeight: 0,
-    boxMaxWidth: 0,
-    verticalDetailsMinWidth: 0,
-});
-const screenWidthRef = ref(1400);
+function toggleGameSelection() {
+    toggableContent.value?.toggle();
+}
 
-const { colors: bulmaColors } = bulmaConstants;
-const starColor = bulmaColors.yellow;
-
-const updateDetailsDirection = preventSequentialCalls(
-    (screenWidth) => {
-        if (!isMobile || !isInitVideoOnly) {
-            isDetailsAtLeft.value = false;
-            isDetailsAtRight.value = false;
-            return;
-        }
-
-        const boxBounds = boxElement.value.getBoundingClientRect();
-        const rightDistance = screenWidth - boxBounds.right;
-        const leftDistance = boxBounds.left;
-        const rightEdgeIsCloser = rightDistance < leftDistance;
-        isDetailsAtRight.value = !rightEdgeIsCloser;
-        isDetailsAtLeft.value = !isDetailsAtRight.value;
-    },
-);
-
-const updateSizeVars = (screenWidth) => {
-    let scale = screenWidth < 1390 ? minScale : 1.0;
-    const isLerping = lerpScaleMinScreen > 0
-        && lerpScaleMaxScreen > 0
-        && minScale !== 1.0;
-
-    if (isLerping) {
-        const widthDiff = lerpScaleMaxScreen - lerpScaleMinScreen;
-        let widthDiffPerc = (screenWidth - lerpScaleMinScreen) / widthDiff;
-        widthDiffPerc = Math.min(1, Math.max(0, widthDiffPerc));
-        scale = minScale + (1 - minScale) * widthDiffPerc;
-    }
-
-    const horizontalVideoSize = {
-        width: BASE_HORIZONTAL_VSIZE.width * scale,
-        height: BASE_HORIZONTAL_VSIZE.height * scale,
-    };
-
-    sizeVars.videoWidth = Math.round(baseVideoSize.width * scale);
-    sizeVars.videoHeight = Math.round(baseVideoSize.height * scale);
-    sizeVars.boxMaxWidth = horizontalVideoSize.width;
-    sizeVars.verticalDetailsMinWidth = horizontalVideoSize.width - sizeVars.videoWidth;
-};
-
-const refreshWindowWidth = () => {
-    if (!window.innerWidth || window.innerWidth <= 0) return;
-    console.log('Screen Size', window.innerWidth);
-    screenWidthRef.value = window.innerWidth;
-};
-
-onBeforeMount(() => {
-    window.addEventListener('resize', refreshWindowWidth);
-    window.addEventListener('load', refreshWindowWidth);
-    window.addEventListener('DOMContentLoaded', refreshWindowWidth);
-    refreshWindowWidth();
-});
-
-onUnmounted(() => {
-    window.removeEventListener('resize', refreshWindowWidth);
-    window.removeEventListener('load', refreshWindowWidth);
-    window.removeEventListener('DOMContentLoaded', refreshWindowWidth);
-});
-
-onMounted(() => {
-    refreshWindowWidth();
-    watch(screenWidthRef, (screenWidth) => {
-        updateDetailsDirection(screenWidth);
-        updateSizeVars(screenWidth);
-    }, { immediate: true });
-});
+function refreshGameSelection() {
+    isSelected.value = toggableContent.value?.isContentOn();
+}
 </script>
 
 <style lang="scss">
+  @use 'sass:color';
   @import '../css/bulma-custom.scss';
 
-  .hover-grow {
-    transition: transform 200ms ease-in-out;
+  // Video Size
+  .game-box {
+    flex-basis: 100%;
 
-    &:hover {
-      transform: scale(1.05);
+    .game-preview {
+      width: 100%;
+      max-width: v-bind('px(baseVideoSize.width)');
+      height: auto;
+      aspect-ratio: v-bind('baseVideoSize.width') / v-bind('baseVideoSize.height');
+      border-radius: calc($box-radius * 1.2);
     }
   }
 
-  .box.game-box {
-    display: inline-block;
-    &.reset-outer-spacing {
-      padding: 0;
-      margin: 0;
+  // Video Header
+  .game-box {
+    .video-header {
+      width: 100%;
+      display: flex;
+      flex-flow: column nowrap;
+      align-items: center;
+
+      .title.game-title {
+        margin-top: 2.5rem;
+        margin-bottom: 0.5rem;
+      }
     }
 
-    &.is-horizontal-video {
-      width: v-bind('px(sizeVars.videoWidth)');
+    &:first-child .video-header .title.game-title {
+      margin-top: 0;
+    }
+  }
+
+  // Game Box Content
+  .game-box {
+    .game-box-content {
+      display: flex;
+      flex-flow: column nowrap;
+      align-items: center;
+    }
+
+    .video-header .title.game-title.page-section-subtitle {
+      padding-top: 0;
+      margin-top: 0;
+    }
+  }
+
+  // Middle Layout
+  .game-box {
+    .game-box-middle, .game-preview-section, .game-preview {
+      width: 100%;
+    }
+
+    .game-box-middle, .game-preview-section {
+      display: flex;
+      flex-flow: column nowrap;
+      align-items: center;
+    }
+  }
+
+  // Game links styling
+  .game-box {
+    .game-links-section {
+      width: 100%;
+      display: flex;
+      flex-flow: column wrap;
+      align-items: center;
     }
 
     &.is-vertical-video {
-      max-width: v-bind('px(sizeVars.boxMaxWidth)');
+      .game-links {
+        max-width: v-bind('px(baseVideoSize.width*1.3)');
+      }
+    }
 
-      .game-box-middle, .game-box-bottom {
-        height: v-bind('px(sizeVars.videoHeight)');
-        min-height: v-bind('px(sizeVars.videoHeight)');
+    &.is-horizontal-video {
+      .game-links {
+        max-width: v-bind('px(baseVideoSize.width*0.6)');
+      }
+    }
+
+    .game-links {
+      display: flex;
+      flex-flow: row wrap;
+      justify-content: space-around;
+      align-items: center;
+      gap: 0.5rem;
+      margin: 0.25rem 0;
+      width: 100%;
+
+      .game-link-wrapper {
+        flex-grow: 1;
+        align-self: center;
+
+        display: flex;
+        flex-flow: row nowrap;
+        justify-content: center;
+        align-items: center;
       }
 
-      .game-box-bottom {
-        min-width: v-bind('px(sizeVars.verticalDetailsMinWidth)');
-      }
+      .game-link {
+        display: inline-block;
+        padding: 0.5rem 1rem;
+        border-radius: $radius;
+        border: $button-border-width solid;
 
-      .game-box-middle {
-        &, video, .video-inner-header, .game-links {
-          width: v-bind('px(sizeVars.videoWidth)');
-          min-width: v-bind('px(sizeVars.videoWidth)');
+        background-color: color.change($dark, $alpha: 0.8);
+        border-color: color.change($white, $alpha: 0.85);
+
+        @at-root .is-mouse &:hover {
+          background-color: color.change($dark, $alpha: 0.95);
+          border-color: color.change($link, $alpha: 1);
         }
       }
+    }
+  }
 
-      .box-content {
+  // Game Sections styling
+  .game-box {
+    .game-sections {
+      display: flex;
+      flex-flow: column nowrap;
+      align-items: stretch;
+
+      .game-section {
+        display: flex;
+        flex-flow: column nowrap;
+        align-items: center;
+      }
+
+      .title.game-section-title {
+        font-size: $size-5;
+        text-align: center;
+        color: $color-title2;
+        margin: 1.5rem 0 0.5rem 0;
+      }
+    }
+
+    .game-details-wrapper {
+      width: 100%;
+    }
+
+    .game-details {
+      .game-details-section {
+        width: 100%;
         display: flex;
         flex-flow: row nowrap;
       }
 
-      .game-box-bottom {
-        .details-columns {
-          flex-flow: column nowrap;
+      .game-details-section-title {
+        text-align: right;
+        font-size: $size-6;
+        color: $color-title3;
+        width: 45%;
+        margin-right: 0.5rem;
+      }
+
+      .game-details-section-content {
+        text-align: left;
+        flex-grow: 3;
+        width: 50%;
+        margin-left: 0.5rem;
+      }
+
+      .game-details-section-title, .game-details-section-content {
+        display: inline-block;
+      }
+    }
+  }
+
+  // Toggable Content Styling
+  .game-box-bottom {
+    margin-top: 0.75rem;
+
+    .game-sections {
+      .game-section:first-child {
+        .title.game-section-title {
+          margin-top: 0;
         }
       }
     }
   }
-</style>
 
-<style lang="scss">
-  @use 'sass:color';
-  @import '../css/bulma-custom.scss';
-  @mixin video-inner-tag-like-colors($back, $back-hover, $border, $border-hover) {
-    border: $button-border-width solid;
-    background-color: color.change($back, $alpha: 0.6);
-    border-color: color.change($border, $alpha: 0.05);
-
-    &:hover {
-      background-color: color.change($back-hover, $alpha: 0.95);
-      border-color: color.change($border-hover, $alpha: 1);
-    }
-  }
-
-  .game-box-middle {
-    position: relative;
-
-    .video-inner-header {
-      display: flex;
-      flex-flow: row nowrap;
-      position: absolute;
-      top: 0;
-      left: 0;
-
-      .game-title.title.title-simple, .video-inner-tag {
-        font-size: $size-6;
-        border-radius: $radius;
-        margin: 0.5rem;
-        padding: 0.5rem ($box-padding - 0.5rem);
-        line-height: inherit;
-      }
-
-      .game-title.title.title-simple {
-        @include video-inner-tag-like-colors($dark, $dark, $white, $white);
-      }
-
-      .video-inner-tag {
-        @include video-inner-tag-like-colors($primary, $primary, $white, $white);
-      }
-    }
-  }
-
-  .game-links {
-    display: flex;
-    flex-flow: row wrap;
-
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    margin-bottom: 1rem;
-    margin-left: 0.25rem;
-
-    .game-link {
-      padding: 0.5rem;
-      border-radius: $radius;
-
-      @include video-inner-tag-like-colors($dark, $dark, $white, $link);
-    }
-  }
-
-  .is-horizontal-video .game-links {
-    justify-content: start;
-    align-items: center;
-    gap: 2rem;
-  }
-
-  .is-vertical-video .game-links {
-    justify-content: space-around;
-    align-items: center;
-    width: 100%;
-  }
-
-  .game-video {
-    width: 100%;
-  }
-
-  .is-horizontal-video .game-video {
-    border-radius: $box-radius $box-radius 0 0;
-  }
-
-  .is-vertical-video .game-video {
-    border-radius: $box-radius 0 0 $box-radius;
-  }
-</style>
-
-<style lang="scss">
-  @use 'sass:color';
-  @import '../css/bulma-custom.scss';
-
-  $base-z-index: 15;
-  $hover-z-index: 30;
-
-  @mixin set-z-indexes($z-index) {
-    z-index: $z-index;
-
-    .video-wrapper {
-      z-index: $z-index;
-
-      .video-poster {
-        z-index: $z-index + 1;
-      }
-
-      .video-overlay {
-        z-index: $z-index + 2;
-      }
-    }
-
-    .game-links, .video-inner-header {
-      z-index: $z-index + 3;
-    }
-
-    .game-box-bottom {
-      z-index: $z-index - 1;
-    }
-  }
-
-  .game-box {
-    @include set-z-indexes($base-z-index);
-  }
-
-  .game-box.is-init-video-only {
-    position: relative;
-
-    .game-box-bottom {
-      position: absolute;
-      opacity: 0;
-      pointer-events: none;
-      transition: opacity 200ms ease-in-out;
-    }
-
-    &.is-horizontal-video .game-box-bottom {
-      left: 0;
-      bottom: 2rem;
-      padding-top: 3rem;
-      transform: translateY(100%);
-    }
-
-    &.is-vertical-video {
-      &.is-details-at-right .game-box-bottom {
-        top: 0;
-        right: 2rem;
-        padding-left: 3rem;
-        transform: translateX(100%);
-      }
-
-      &.is-details-at-left .game-box-bottom {
-        top: 0;
-        left: 2rem;
-        padding-right: 3rem;
-        transform: translateX(-100%);
-      }
-    }
-
-    &:hover {
-      @include set-z-indexes($hover-z-index);
-
-      .game-box-bottom {
-        transition: opacity 200ms ease-in-out;
-        opacity: 1;
-        pointer-events: all;
-        margin: 0;
-      }
-    }
-
-    .game-video {
-      border-radius: $box-radius;
-    }
-
-    .game-box-bottom {
-      border-radius: $box-radius;
-      background-color: color.change(#1c4b52, $alpha: 0.95);
-    }
-  }
-
-  .game-box-bottom {
-    .details-columns {
-      display: flex;
-      align-items: start;
-    }
-  }
-
-  .is-horizontal-video .game-box-bottom {
-    .details-columns {
-      flex-flow: row nowrap;
-
-      .highlights-column {
-        flex: 5;
-      }
-
-      .bullets-column {
-        flex: 4;
-      }
-    }
-  }
-
-  .is-vertical-video .game-box-bottom {
-    .details-columns {
-      flex-flow: column nowrap;
-    }
-  }
-
-  .game-box-bottom {
-    padding: 1rem $box-padding;
-    .title {
-      color: $link;
-    }
-
-    &.box {
-      display: inline-block;
-    }
+  // Game Highlights Styling
+  .game-highlights {
+    font-size: $size-5;
 
     .bullet-list {
-      display: block;
+      margin: 0.5rem 1rem;
+      gap: 1.5rem;
+    }
+  }
+
+  // Game Integrations Styling
+  .game-integrations {
+    .inline-list {
+      margin: 0 1rem;
+      justify-content: center;
+      align-items: center;
+    }
+  }
+
+  // Semantic Hide
+  .game-box {
+    .game-links-section .title,
+    .game-preview-section .title
+    {
+      @include semantic-hide;
+    }
+  }
+
+  // Gallery View Layout
+  @include from($game-gallery-breakpoint) {
+    .game-box .game-links .game-link {
+      padding: 0.25rem 0.5rem;
     }
 
-    .game-highlights {
-      .bullet-list-item {
-        margin-top: 0.75rem;
-        margin-bottom: 1rem;
-      }
+    .game-box .video-header .title.game-title.page-section-subtitle {
+      margin-top: 0rem;
+    }
+  }
+  @include tablet {
+    .game-box {
+      flex-basis: 49%;
+      flex-grow: 1;
+      max-width: v-bind('px(baseVideoSize.width)');
+    }
+  }
 
-      .bullet-list, .bullet-list-item {
-        &, span {
-          font-size: $size-5;
+  @include desktop {
+    .game-box {
+      flex-basis: 40%;
+
+      &.is-vertical-video {
+        flex-basis: v-bind('px(baseVideoSize.width * 0.4)');
+      }
+    }
+  }
+
+  // Game box Selected
+  .game-box {
+    .game-box-content {
+      // margin-top: 2rem;
+      padding: 1rem 0.25rem;
+      border-radius: $box-radius;
+      transition: background-color 200ms ease-out;
+    }
+
+    .game-preview-section {
+      border: solid 1px rgba(0, 0, 0, 0);
+      border-radius: $box-radius;
+      overflow: hidden;
+      transition: border 200ms ease-out;
+
+      .game-preview {
+        transition: transform 200ms ease-out;
+      }
+    }
+
+    .game-box-content {
+      transition: background-color 200ms ease-out;
+    }
+  }
+
+  // Gamebox Hover
+  .game-box {
+    @at-root .is-mouse &:hover, &.is-selected {
+      .game-box-content {
+        cursor: pointer;
+        background-color: $game-box-selected-color;
+
+        .game-preview-section {
+          border: solid 1px color.change($link, $alpha: 0.3);
+
+          .game-preview {
+            position: relative;
+            transform: scale(1.25);
+          }
         }
       }
     }
 
-    .game-integrations {
-      margin-bottom: 0.5rem;
+    @include from($game-gallery-breakpoint) {
+      &.is-selected .game-box-content {
+        background-color: $game-box-selected-dark-color;
+      }
+    }
+  }
+
+  // Gallery View Styling
+  .screen-overlay {
+    cursor: pointer;
+    position: fixed;
+    z-index: -9999;
+    width: 100%;
+    height: 100%;
+    top: 0;
+    left: 0;
+    opacity: 0;
+    background-color: black;
+    transition: opacity 200ms ease-out;
+  }
+
+  .preview-info-button.button {
+    display: none;
+  }
+
+  @include from($game-gallery-breakpoint) {
+    .game-box {
+      // Set screen overlay
+      &.is-selected .screen-overlay {
+        opacity: 0.95;
+      }
+
+      .game-box-content {
+        position: relative;
+        overflow: visible;
+      }
+
+      // Info icon inside video
+      .game-preview-section {
+        position: relative;
+      }
+
+      .preview-info-button {
+        position: absolute;
+        z-index: 5;
+        bottom: 0;
+        right: 0;
+
+        display: flex;
+        flex-flow: column nowrap;
+        align-items: center;
+        justify-items: center;
+        text-align: center;
+        width: 2rem;
+        height: 2rem;
+        margin: 0.5em;
+      }
+
+      .preview-info-button.button.is-link.is-outlined {
+        background-color: color.change($link, $alpha: 0.2);
+      }
+      @at-root .is-mouse &:hover, &.is-selected {
+        .preview-info-button.button.is-link.is-outlined {
+          background-color: $link;
+          color: $white;
+        }
+      }
+
+      // Set z-indexes
+      $z-index: 500;
+      &.is-selected {
+        .game-box-content {
+          z-index: $z-index;
+        }
+
+        .video-wrapper {
+          z-index: $z-index;
+
+          .video-poster {
+            z-index: $z-index + 1;
+          }
+
+          .video-overlay {
+            z-index: $z-index + 2;
+          }
+        }
+
+        .preview-info-button {
+          z-index: $z-index + 3;
+        }
+
+        .toggable-content .inner-content {
+          z-index: $z-index - 1;
+        }
+
+        .screen-overlay {
+          z-index: $z-index - 2;
+        }
+      }
     }
   }
 </style>
